Don't call wcslen() on char* string
getenv() gets us ASCII string. Reinterpret casting it to wide char and
calling wcslen() can end on a different NUL character in the env
variables block, depending if the lenght of PATH is even or odd.

Calling strlen() is the right thing to do to get the real length,
std::wstring constructor know how to deal with ASCII strings.

This fixes eldenring.exe not starting due to not finding lua.dll on some
systems after hooking.